Fix recasting bug, add floating point power level
- Fix the bug that the reel doesn't close after recasting in bottom
  fishing, this is caused by the removal of click() in tackle.cast()

- Change the casting power level from integer to floating point number

- Update README.md

+++ src/userprofile.py
@@ -22,7 +22,7 @@ class UserProfile():
                  base_iteration: int,
                  # bottom
                  check_delay: float,
-                 cast_power_level: int,
+                 cast_power_level: float,
                  # marine
                  pirk_duration: float,
                  pirk_delay: float,
